+++ projects/pysan/fuzz_ansible_runner.py
@@ -17,7 +17,7 @@
 import os
 import sys
 import atheris
-import sanlib
+import pysan
 
 import pexpect
 import ansible_runner
@@ -60,6 +60,7 @@ def TestOneInput(data):
 
 
 def main():
+    pysan.pysan_add_hooks()
     atheris.instrument_all()
     atheris.Setup(sys.argv, TestOneInput, enable_python_coverage=True)
     atheris.Fuzz()
