+++ projects/pysan/fuzz_libvcs.py
@@ -16,7 +16,7 @@
 import os
 import sys
 import atheris
-import sanlib
+import pysan
 
 import libvcs
 from libvcs.shortcuts import create_repo, create_repo_from_pip_url
@@ -38,6 +38,7 @@ def TestOneInput(data):
 
 
 def main():
+    pysan.pysan_add_hooks()
     atheris.instrument_all()
     atheris.Setup(sys.argv, TestOneInput, enable_python_coverage=True)
     atheris.Fuzz()
